dc - fixed eslint and put tests into one file
