+++ frontend/src/main/pages/GeneralEducation/Search/GeneralEducationSearchPage.jsx
